Optimize ServicesHomeScreen rendering
